+++ src/gamspy/_symbols/equation.py
@@ -319,7 +319,7 @@ class Equation(gt.Equation, operable.Operable, Symbol):
             if records is not None:
                 self.setRecords(records, uels_on_axes=uels_on_axes)
             else:
-                self.container._run()
+                self.container._synch_with_gams()
 
             container.miro_protect = previous_state
 
@@ -347,7 +347,7 @@ class Equation(gt.Equation, operable.Operable, Symbol):
         self._set_definition(domain, rhs)
         self._is_dirty = True
 
-        self.container._run()
+        self.container._synch_with_gams()
 
     def __eq__(self, other):  # type: ignore
         return expression.Expression(self, "=e=", other)
@@ -597,7 +597,7 @@ class Equation(gt.Equation, operable.Operable, Symbol):
 
     def setRecords(self, records: Any, uels_on_axes: bool = False) -> None:
         super().setRecords(records, uels_on_axes)
-        self.container._run()
+        self.container._synch_with_gams()
 
     @property
     def type(self):