+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 GAMSPy CHANGELOG
 ================
 
+GAMSPy 0.13.1
+-------------
+- General
+  - Support output redirection for NEOS backend.
+- Documentation
+  - Update model documentation to show how to redirect NEOS output to a file.
+
 GAMSPy 0.13.0
 -------------
 - General